This is the interface the user interacts with.
The sensor readings are shown on the bottom of the screen. Every 30 minutes the, global time that the readings were taken is shown in the top left. The current battery level is found in the top right. The L in the top right represent a successful transmission to the LoRa base station. The X represents no Bluetooth device requested a data pull. This changes to a B if a pull was processed.